pyproject.toml
[project]
name = "bookworm"
version = "0.1.0"
description = "A Django REST API for a book library"
readme = "README.md"
requires-python = ">=3.12"
# ==== isort ====

classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "django>=5.1.6",
  "django-filter>=25.1",
  "djangorestframework>=3.15.2",
  "djlint>=1.36.4",
  "drf-spectacular>=0.28.0",
  "factory-boy>=3.3.3",
  "pillow>=11.1.0",
  "pre-commit>=4.1.0",
  "psycopg2-binary>=2.9.10",
  "python-dotenv>=1.0.1",
  "tox>=4.24.1",
  "tox-uv>=1.25.0",
]

[tool.isort]
profile = "black"
line_length = 119
known_first_party = [
  "tests",
  "scripts",
  "hooks",
]

[tool.pyproject-fmt]
keep_full_version = true

# ==== djLint ====

[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002,T028"
ignore_blocks = "raw"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "jinja"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
